IMAGE_RECOGNITION/requirements.txt
fastapi
uvicorn
google-cloud-vision
python-multipart
pillow
python-dotenv
